requirements.txt
Django
psycopg2
psycopg2-binary
python-decouple
pillow
djangorestframework
djangorestframework-simplejwt==5.3.1
django-cors-headers
requests
gunicorn
openpyxl
pandas
django-redis
django-filter
channels 
channels-redis
